add codification script
accepts keyboard keys f,d,s,j,k,l as input, maps each one of them correspondigly(1,2,3,4,5,6), creates a numerical key in the form of '123456' and uses that as a key in a dictionary to print the corresponding braille character. Space is used to re enter the numkey generation process. Escape terminates the script. There's also a dictionary that uses braille characters as keys with their corresponding numkeys as values.